+++ src/components/shared/transactionInfo/index.js
@@ -3,170 +3,15 @@ import withData from '@utils/withData';
 import { withTranslation } from 'react-i18next';
 import { getAccounts } from '@api/account';
 import AccountVisual from '@toolbox/accountVisual';
-import AccountMigration from '@shared/accountMigration';
 import LiskAmount from '@shared/liskAmount';
-import VoteItem from '@shared/voteItem';
 import { tokenMap, MODULE_ASSETS_NAME_ID_MAP } from '@constants';
-import MultiSignatureReview from '@shared/multiSignatureReview';
-
+import VoteDelegate from './voteDelegate';
+import Reclaim from './reclaim';
+import Send from './send';
+import RegisterDelegate from './registerDelegate';
+import UnlockBalance from './unlockBalance';
+import RegisterMultisignatureGroup from './registerMultisignatureGroup';
 import styles from './transactionInfo.css';
-
-const ItemList = ({ items, heading }) => (
-  <div className={styles.contentItem}>
-    <span className={styles.contentHeading}>{heading}</span>
-    <div className={styles.voteItems}>
-      {Object.keys(items).map(address => (
-        <VoteItem
-          key={`vote-item-${address}`}
-          address={address}
-          vote={items[address]}
-          title={items[address].username}
-        />
-      ))}
-    </div>
-  </div>
-);
-
-const InfoColumn = ({ title, children, className }) => (
-  <div className={`${styles.infoColumn} ${className}`}>
-    <span className={styles.infoTitle}>{title}</span>
-    <span className={styles.infoValue}>
-      {children}
-    </span>
-  </div>
-);
-
-const VoteDelegate = ({
-  added, edited, removed, fee, t,
-}) => {
-  const addedLength = Object.keys(added).length;
-  const editedLength = Object.keys(edited).length;
-  const removedLength = Object.keys(removed).length;
-
-  return (
-    <>
-      {addedLength ? <ItemList heading={t('Added votes')} items={added} /> : null}
-      {editedLength ? <ItemList heading={t('Changed votes')} items={edited} /> : null}
-      {removedLength ? <ItemList heading={t('Removed votes')} items={removed} /> : null}
-      <div className={styles.infoContainer}>
-        <InfoColumn title={t('Total votes after confirmation')} className="total-votes">{`${addedLength + editedLength}/10`}</InfoColumn>
-      </div>
-    </>
-  );
-};
-
-const RegisterMultisignatureGroup = ({
-  t, members, fee, numberOfSignatures,
-}) => (
-  <MultiSignatureReview t={t} members={members} fee={fee} numberOfSignatures={numberOfSignatures} />
-);
-
-const Reclaim = ({ account, t }) => (
-  <>
-    <section>
-      <AccountMigration account={account.info.LSK} showBalance={false} />
-    </section>
-    <section>
-      <label>{t('Balance to reclaim')}</label>
-      <LiskAmount
-        val={Number(account.info.LSK.legacy.balance)}
-        token={tokenMap.LSK.key}
-      />
-    </section>
-  </>
-);
-
-const Send = ({
-  fields, token, transaction = {}, t,
-}) => (
-  <>
-    <section>
-      <label>{t('Recipient')}</label>
-      <label className="recipient-value">
-        <AccountVisual address={fields.recipient.address} size={25} />
-        <label className={`${styles.information} recipient-confirm`}>
-          {fields.recipient.title || fields.recipient.address}
-        </label>
-        { fields.recipient.title ? (
-          <span className={styles.secondText}>
-            {fields.recipient.address}
-          </span>
-        ) : null }
-      </label>
-    </section>
-    <section className={styles.msignRow}>
-      <div className={styles.col}>
-        <label>{t('Transaction ID')}</label>
-        <label>
-          {transaction.id ? Buffer.from(transaction.id, 'hex') : '-'}
-        </label>
-      </div>
-      <div className={styles.col}>
-        <label>{t('Amount')}</label>
-        <label>
-          <LiskAmount
-            val={transaction.asset?.amount}
-            token={token}
-          />
-        </label>
-      </div>
-    </section>
-    <section>
-      <label>{t('Message')}</label>
-      <label className="recipient-value">
-        {transaction.asset?.data}
-      </label>
-    </section>
-  </>
-);
-
-const RegisterDelegate = ({ account, nickname, t }) => (
-  <section className="summary-container">
-    <label className="nickname-label">{t('Your nickname')}</label>
-    <div className={styles.userInformation}>
-      <AccountVisual
-        className={styles.accountVisual}
-        address={account.summary?.address}
-        size={25}
-      />
-      <span className={`${styles.nickname} nickname`}>{nickname}</span>
-      <span className={`${styles.address} address`}>{account.summary?.address}</span>
-    </div>
-  </section>
-);
-
-const UnlockBalance = ({ account, t, transaction = {} }) => (
-  <>
-    <section>
-      <label>{t('Sender')}</label>
-      <label>
-        <AccountVisual address={account.summary.address} size={25} />
-        <label>
-          {account.summary.address}
-        </label>
-      </label>
-    </section>
-    <section className={styles.msignRow}>
-      <div className={styles.col}>
-        <label>{t('Transaction ID')}</label>
-        <label>
-          {transaction.id ? Buffer.from(transaction.id, 'hex') : '-'}
-        </label>
-      </div>
-      <div className={styles.col}>
-        <label>{t('Amount to unlock')}</label>
-        <label>
-          <LiskAmount
-            val={transaction.asset.unlockObjects.reduce(
-              (total, { amount }) => total + Number(amount), 0,
-            )}
-            token={tokenMap.LSK.key}
-          />
-        </label>
-      </div>
-    </section>
-  </>
-);
 
 const CustomTransactionInfo = ({ moduleAssetId, ...restProps }) => {
   switch (moduleAssetId) {
